Make accident_count_by_severity_widget.py work also for street location.
