+++ anyway/widgets/all_locations_widgets/__init__.py
@@ -1,0 +1,1 @@
+from . import accident_count_by_severity_widget
